+++ benchmark/src/org/pmw/benchmark/SingleThreadBenchmarkRunner.java
@@ -22,8 +22,8 @@ public class SingleThreadBenchmarkRunner extends AbstractBenchmarkRunner {
 	}
 
 	@Override
-	protected final int countLogEntries() {
-		return LOGGING_ITERATIONS * 5;
+	protected final long countLogEntries() {
+		return LOGGING_ITERATIONS * 5L;
 	}
 
 	@Override
